Exception handling

+++ src/main/java/com/devsuperior/services/ClientService.java
@@ -1,10 +1,11 @@
 package com.devsuperior.services;
 
-import java.util.NoSuchElementException;
-import java.util.Optional;
-
-import com.devsuperior.Dto.ClientDto;
+import com.devsuperior.dto.ClientDto;
+import com.devsuperior.services.exceptions.DatabaseException;
+import com.devsuperior.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.modelmapper.ModelMapper;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,7 @@ import com.devsuperior.entities.Client;
 import com.devsuperior.repositories.ClientRepository;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -24,8 +26,9 @@ public class ClientService {
 
 	@Transactional(readOnly = true)
 	public ClientDto findById(Long id) {
-		Optional<Client> clientOptional = clientRepository.findById(id);
-		Client client = clientOptional.orElseThrow(NoSuchElementException::new);
+		Client client = clientRepository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado")
+		);
         return modelMapper.map(client, ClientDto.class);
 	}
 
@@ -35,13 +38,43 @@ public class ClientService {
         return clients.map(client -> modelMapper.map(client, ClientDto.class));
 	}
 
+	@Transactional
 	public ClientDto insert(ClientDto clientDto) {
 		Client client = modelMapper.map(clientDto, Client.class);
 		Client clientSaved = clientRepository.save(client);
 		return modelMapper.map(clientSaved, ClientDto.class);
 	}
 
-	public void deleteById(Long id) {
-		clientRepository.deleteById(id);
+	@Transactional
+	public ClientDto update(Long id, ClientDto clientDto) {
+		try {
+			Client client = clientRepository.getReferenceById(id);
+			copyDtoToEntity(clientDto, client);
+			Client clientUpdated = clientRepository.save(client);
+			return modelMapper.map(clientUpdated, ClientDto.class);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+	}
+
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if (!clientRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+			clientRepository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
+	}
+
+	private static void copyDtoToEntity(ClientDto clientDto, Client client) {
+		client.setName(clientDto.getName());
+		client.setChildren(clientDto.getChildren());
+		client.setCpf(clientDto.getCpf());
+		client.setIncome(clientDto.getIncome());
+		client.setBirthDate(clientDto.getBirthDate());
 	}
 }
